+++ src/runtime/node/node_process.rs
@@ -49,9 +49,9 @@ pub(crate) extern "C" fn get_exec_argv(global: &JSGlobalObject) -> JSValue {
 
 // @190n: this may need to be noreturn
 #[unsafe(export_name = "Bun__Process__exit")]
-pub extern "C" fn exit(global_object: &JSGlobalObject, code: u8) {
+pub extern "C" fn exit(global_object: &JSGlobalObject, code: i32) {
     let vm = global_object.bun_vm().as_mut();
-    vm.exit_handler.exit_code = i32::from(code);
+    vm.exit_handler.exit_code = code;
     if let Some(worker) = vm.worker_ref() {
         // @190n: we may need to use requestTerminate or throwTerminationException
         // instead to terminate the worker sooner
